Cloud_vertex_ai/requirements.txt
pandas==2.2.3
scikit-learn==1.6.0
numpy==1.26.4
gcsfs==0.6.2
joblib==1.4.2
google-cloud-storage==2.19.0